refactor(ops): mutualise dotenv + connexion DB dans lib/env et lib/db
createOpsDb() + bootstrap env partagés remplacent le bloc connexion copié dans 6 scripts.
