+++ scripts/ops/audit/audit-parcours-ds-integrity.ts
@@ -20,16 +20,10 @@
  * Prérequis : .env.local avec DATABASE_URL + DEMARCHES_SIMPLIFIEES_*
  */
 
-import { config } from "dotenv";
-config({ path: ".env.local" });
-config({ path: ".env" });
-
 import { writeFileSync } from "node:fs";
-import { createRedactor } from "../lib/anonymize";
-import { drizzle } from "drizzle-orm/postgres-js";
-import postgres from "postgres";
 import { and, eq, inArray, desc } from "drizzle-orm";
-import * as schema from "@/shared/database/schema";
+import { createOpsDb } from "../lib/db";
+import { createRedactor } from "../lib/anonymize";
 import {
   parcoursPrevention,
   users,
@@ -79,15 +73,7 @@ for (const step of [Step.ELIGIBILITE, Step.DIAGNOSTIC, Step.DEVIS, Step.FACTURES
 }
 
 // --- DB ---
-const connectionString =
-  process.env.DATABASE_URL ??
-  (() => {
-    const { DB_HOST, DB_PORT, DB_USER, DB_PASSWORD, DB_NAME } = process.env;
-    if (!DB_HOST) throw new Error("DATABASE_URL ou DB_HOST requis");
-    return `postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}`;
-  })();
-const client = postgres(connectionString, { max: 5, idle_timeout: 10 });
-const db = drizzle(client, { schema });
+const { db, client } = createOpsDb();
 
 // --- Types ---
 interface DsSearchHit {
